requirements.txt
pandas
llama-index==0.14.13
numpy==2.4.1
llama-index-retrievers-bm25==0.6.5
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-groq==0.4.1
langchain_community==0.3.29
langchain==0.3.27
langchain-core==0.3.83
langchain-groq==0.3.8
openpyxl==3.1.5
pgvector==0.4.2 
psycopg2-binary==2.9.11
llama-index-vector-stores-postgres==0.7.3 
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlalchemy
asyncpg
PyYAML
sqlmodel
requests
httpx